# api/contracts/control-plane.openapi.json
{
  "components": {
    "schemas": {
...
      "BacktestClaimRequest": {
        "additionalProperties": false,
        "properties": {
          "executionName": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executionname"
          }
        },
        "title": "BacktestClaimRequest",
        "type": "object"
      },
      "BacktestCompleteRequest": {
        "additionalProperties": false,
        "properties": {
          "artifactManifestPath": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifactmanifestpath"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          }
        },
        "title": "BacktestCompleteRequest",
        "type": "object"
      },
      "BacktestFailRequest": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "BacktestFailRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Asset Allocation API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/internal/backtests/runs/claim": {
      "post": {
        "operationId": "claim_backtest_run_api_internal_backtests_runs_claim_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestClaimRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Claim Backtest Run Api Internal Backtests Runs Claim Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Claim Backtest Run",
        "tags": [
          "Internal"
        ]
      }
    },
    "/api/internal/backtests/runs/{run_id}": {
      "get": {
        "operationId": "get_backtest_run_api_internal_backtests_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Backtest Run Api Internal Backtests Runs  Run Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Backtest Run",
        "tags": [
          "Internal"
        ]
      }
    },
    "/api/internal/backtests/runs/{run_id}/complete": {
      "post": {
        "operationId": "complete_backtest_run_api_internal_backtests_runs__run_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Complete Backtest Run Api Internal Backtests Runs  Run Id  Complete Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Backtest Run",
        "tags": [
          "Internal"
        ]
      }
    },
    "/api/internal/backtests/runs/{run_id}/fail": {
      "post": {
        "operationId": "fail_backtest_run_api_internal_backtests_runs__run_id__fail_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestFailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Fail Backtest Run Api Internal Backtests Runs  Run Id  Fail Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fail Backtest Run",
        "tags": [
          "Internal"
        ]
      }
    },
    "/api/internal/backtests/runs/{run_id}/heartbeat": {
      "post": {
        "operationId": "backtest_heartbeat_api_internal_backtests_runs__run_id__heartbeat_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Backtest Heartbeat Api Internal Backtests Runs  Run Id  Heartbeat Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Backtest Heartbeat",
        "tags": [
          "Internal"
        ]
      }
    },
    "/api/internal/backtests/runs/{run_id}/start": {
      "post": {
        "operationId": "start_backtest_run_api_internal_backtests_runs__run_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Start Backtest Run Api Internal Backtests Runs  Run Id  Start Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Backtest Run",
        "tags": [
          "Internal"
        ]
      }
    },
...
    "/api/internal/rankings/{name}": {
      "get": {
        "operationId": "get_ranking_schema_api_internal_rankings__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Ranking Schema Api Internal Rankings  Name  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ranking Schema",
        "tags": [
          "Internal"
        ]
      }
    },
    "/api/internal/rankings/{name}/revision": {
      "get": {
        "operationId": "get_ranking_schema_revision_api_internal_rankings__name__revision_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Ranking Schema Revision Api Internal Rankings  Name  Revision Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ranking Schema Revision",
        "tags": [
          "Internal"
        ]
      }
    },
    "/api/internal/regimes/current": {
      "get": {
        "operationId": "get_regime_current_api_internal_regimes_current_get",
        "parameters": [
          {
            "in": "query",
            "name": "modelName",
            "required": false,
            "schema": {
              "default": "default-regime",
              "minLength": 1,
              "title": "Modelname",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "modelVersion",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Modelversion"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Regime Current Api Internal Regimes Current Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Regime Current",
        "tags": [
          "Internal"
        ]
      }
    },
...
    "/api/internal/regimes/models/{name}/active": {
      "get": {
        "operationId": "get_active_regime_model_api_internal_regimes_models__name__active_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Active Regime Model Api Internal Regimes Models  Name  Active Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Active Regime Model",
        "tags": [
          "Internal"
        ]
      }
    },
    "/api/internal/regimes/models/{name}/revision": {
      "get": {
        "operationId": "get_regime_model_revision_api_internal_regimes_models__name__revision_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Regime Model Revision Api Internal Regimes Models  Name  Revision Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Regime Model Revision",
        "tags": [
          "Internal"
        ]
      }
    },
...
    "/api/internal/strategies/{name}": {
      "get": {
        "operationId": "get_strategy_api_internal_strategies__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Strategy Api Internal Strategies  Name  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Strategy",
        "tags": [
          "Internal"
        ]
      }
    },
    "/api/internal/strategies/{name}/revision": {
      "get": {
        "operationId": "get_strategy_revision_api_internal_strategies__name__revision_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Strategy Revision Api Internal Strategies  Name  Revision Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Strategy Revision",
        "tags": [
          "Internal"
        ]
      }
    },
    "/api/internal/universes/{name}": {
      "get": {
        "operationId": "get_universe_config_api_internal_universes__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Universe Config Api Internal Universes  Name  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Universe Config",
        "tags": [
          "Internal"
        ]
      }
    },
    "/api/internal/universes/{name}/revision": {
      "get": {
        "operationId": "get_universe_config_revision_api_internal_universes__name__revision_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Universe Config Revision Api Internal Universes  Name  Revision Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Universe Config Revision",
        "tags": [
          "Internal"
        ]
      }
    },
...
    "/api/rankings/": {
      "get": {
        "operationId": "list_ranking_schemas_api_rankings__get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RankingSchemaSummaryResponse"
                  },
                  "title": "Response List Ranking Schemas Api Rankings  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Ranking Schemas",
        "tags": [
          "Rankings"
        ]
      },
      "post": {
        "operationId": "save_ranking_schema_api_rankings__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RankingSchemaUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Save Ranking Schema Api Rankings  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Ranking Schema",
        "tags": [
          "Rankings"
        ]
      }
    },
...
    "/api/rankings/{name}": {
      "delete": {
        "operationId": "delete_ranking_schema_api_rankings__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Ranking Schema Api Rankings  Name  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Ranking Schema",
        "tags": [
          "Rankings"
        ]
      }
    },
...
    "/api/regimes/current": {
      "get": {
        "operationId": "get_current_regime_api_regimes_current_get",
        "parameters": [
          {
            "in": "query",
            "name": "modelName",
            "required": false,
            "schema": {
              "default": "default-regime",
              "minLength": 1,
              "title": "Modelname",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "modelVersion",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Modelversion"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Current Regime Api Regimes Current Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Current Regime",
        "tags": [
          "Regimes"
        ]
      }
    },
...
  }
}